LayoutTests/inspector/dom-debugger/url-breakpoints-fetch-expected.txt

Tests for URL breakpoints.


== Running test suite: URLBreakpoint
-- Running test case: URLBreakpoint.ShouldNotPauseOnDisabledBreakpoint.Fetch
Adding "text:data" URL Breakpoint...
Disabling breakpoint...
Request Fetch "resources/dataFetch.json"...
PASS: Should not pause for disabled breakpoint.
-- Running test teardown.

-- Running test case: URLBreakpoint.ShouldNotPauseOnNotContainingURL.Fetch
Adding "text:nonexistant-url" URL Breakpoint...
Request Fetch "resources/dataFetch.json"...
PASS: Should not pause for breakpoint with different URL.
-- Running test teardown.

-- Running test case: URLBreakpoint.RemoveBreakpoint.Fetch
Adding "text:data" URL Breakpoint...
Removing breakpoint...
Request Fetch "resources/dataFetch.json"...
PASS: Should not pause for removed breakpoint.
-- Running test teardown.

-- Running test case: URLBreakpoint.RemoveDisabledBreakpoint.Fetch
Adding "text:data" URL Breakpoint...
Disabling breakpoint...
Removing breakpoint...
Request Fetch "resources/dataFetch.json"...
PASS: Should not pause for removed disabled breakpoint.
-- Running test teardown.
